@@ -25,6 +25,7 @@
     "@babel/preset-env": "^7.15.0",
     "@babel/preset-react": "^7.14.5",
     "@babel/register": "^7.15.3",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.2.2",
     "babel-regenerator-runtime": "^6.5.0",
@@ -32,8 +33,8 @@
     "copy-webpack-plugin": "^9.0.1",
     "css-loader": "^6.2.0",
     "dev-tools": "^0.1.6",
-    "enzyme": "^2.8.2",
-    "enzyme-to-json": "^1.5.1",
+    "enzyme": "^3.11.0",
+    "enzyme-to-json": "^3.6.2",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-loader": "^4.0.2",
@@ -51,14 +52,14 @@
     "npm-force-resolutions": "0.0.10",
     "prettier": "^1.19.1",
     "prop-types": "^15.5.8",
-    "react-test-renderer": "^15.5.4",
+    "react-test-renderer": "^17.0.2",
     "regenerator-runtime": "^0.10.5",
     "replace-in-file": "^3.4.4",
     "request": "^2.88.2",
     "sass-loader": "^12.1.0",
     "style-loader": "^3.2.1",
     "url-loader": "^4.1.1",
-    "webpack": "^5.51.1",
+    "webpack": "^5.52.0",
     "webpack-cli": "^4.8.0",
     "webpack-dev-server": "^4.0.0"
   },
@@ -91,9 +92,9 @@
     "minimongo": "^6.9.0",
     "onchange": "^6.1.0",
     "path": "^0.12.7",
-    "react": "^15.6.2",
-    "react-dom": "^15.6.2",
-    "react-redux": "^5.1.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.5",
     "react-router": "^3.2.6",
     "react-router-redux": "^4.0.8",
     "react-transition-group": "^2.9.0",
