@@ -39,17 +39,14 @@ export class CollectionsController extends Component {
             isFetchingCollections: false
         };
 
-        this.handleCollectionSelection = this.handleCollectionSelection.bind(this);
-        this.handleCollectionCreateSuccess = this.handleCollectionCreateSuccess.bind(this);
-
         this.isViewer = this.props.user.userType === "VIEWER";
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
         return this.fetchCollections();
     }
 
-    componentWillReceiveProps(nextProps) {
+    UNSAFE_componentWillReceiveProps(nextProps) {
         // CollectionsController handles removing any collections from allCollections state
         // having an ID in the toDelete object means that this component needs to remove it from state
         // this stops other components having to understand and handle allCollections state.
@@ -167,7 +164,7 @@ export class CollectionsController extends Component {
         }
     }
 
-    handleCollectionCreateSuccess(newCollection) {
+    handleCollectionCreateSuccess = newCollection => {
         const mappedCollection = collectionMapper.collectionResponseToState(newCollection);
         let collections = [...this.props.collections, mappedCollection];
         collections.sort((collection1, collection2) => {
@@ -185,9 +182,9 @@ export class CollectionsController extends Component {
         this.props.dispatch(push(`${this.props.rootPath}/collections/${mappedCollection.id}`));
         this.fetchCollections();
         document.getElementById(mappedCollection.id).scrollIntoView();
-    }
+    };
 
-    handleCollectionSelection(collection) {
+    handleCollectionSelection = collection => {
         if (this.isViewer) {
             cookies.add("collection", collection.id, null);
             this.props.dispatch(updateWorkingOn(collection.id, collection.name));
@@ -195,7 +192,7 @@ export class CollectionsController extends Component {
             return;
         }
         this.props.dispatch(push(`${this.props.rootPath}/collections/${collection.id}`));
-    }
+    };
 
     render() {
         return (
