@@ -80,21 +80,9 @@ export class CollectionDetailsController extends Component {
             drawerIsAnimatable: false,
             drawerIsVisible: false
         };
-
-        this.handleDrawerTransitionEnd = this.handleDrawerTransitionEnd.bind(this);
-        this.handleDrawerCloseClick = this.handleDrawerCloseClick.bind(this);
-        this.handleCollectionDeleteClick = this.handleCollectionDeleteClick.bind(this);
-        this.handleCollectionApproveClick = this.handleCollectionApproveClick.bind(this);
-        this.handleCollectionPageClick = this.handleCollectionPageClick.bind(this);
-        this.handleCollectionPageEditClick = this.handleCollectionPageEditClick.bind(this);
-        this.handleCollectionPageDeleteClick = this.handleCollectionPageDeleteClick.bind(this);
-        this.handleCancelPageDeleteClick = this.handleCancelPageDeleteClick.bind(this);
-        this.handleRestoreDeletedContentClose = this.handleRestoreDeletedContentClose.bind(this);
-        this.handleRestoreMultiDeletedContentSuccess = this.handleRestoreMultiDeletedContentSuccess.bind(this);
-        this.handleRestoreSingleDeletedContentSuccess = this.handleRestoreSingleDeletedContentSuccess.bind(this);
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
         if (!auth.canViewCollectionsDetails(this.props.user)) {
             this.props.dispatch(push(`${this.props.rootPath}/collections`));
             return;
@@ -109,7 +97,7 @@ export class CollectionDetailsController extends Component {
         }
     }
 
-    componentWillReceiveProps(nextProps) {
+    UNSAFE_componentWillReceiveProps(nextProps) {
         // Open and close edit collection modal
         if (nextProps.routes[nextProps.routes.length - 1].path === "edit") {
             this.setState({ isEditingCollection: true });
@@ -219,7 +207,7 @@ export class CollectionDetailsController extends Component {
         this.props.dispatch(emptyWorkingOn());
     }
 
-    handleCollectionDeleteClick(collectionID) {
+    handleCollectionDeleteClick = collectionID => {
         this.props.dispatch(push(`${this.props.rootPath}/collections`));
         collections
             .delete(collectionID)
@@ -241,9 +229,9 @@ export class CollectionDetailsController extends Component {
                 console.error(`Error deleting collection '${collectionID}'`, error);
                 collectionDetailsErrorNotifications.deleteCollection(error);
             });
-    }
+    };
 
-    handleCollectionApproveClick() {
+    handleCollectionApproveClick = () => {
         const activeCollection = this.props.activeCollection;
         const collectionID = this.props.collectionID;
         if (!collectionMapper.collectionCanBeApproved(activeCollection)) {
@@ -283,9 +271,9 @@ export class CollectionDetailsController extends Component {
                 console.error("Error approving collection", error);
                 collectionDetailsErrorNotifications.approveCollection(error);
             });
-    }
+    };
 
-    handleCancelPageDeleteClick(uri) {
+    handleCancelPageDeleteClick = uri => {
         if (!uri) {
             notifications.add({
                 type: "warning",
@@ -339,9 +327,9 @@ export class CollectionDetailsController extends Component {
                 collectionDetailsErrorNotifications.cancelPageDelete(error, uri, this.props.collectionID);
                 console.error(`Error removing pending delete of page '${uri}' from collection '${this.props.collectionID}'`, error);
             });
-    }
+    };
 
-    handleDrawerTransitionEnd() {
+    handleDrawerTransitionEnd = () => {
         this.setState({
             drawerIsAnimatable: false
         });
@@ -353,9 +341,9 @@ export class CollectionDetailsController extends Component {
             this.props.dispatch(emptyActiveCollection());
             this.props.dispatch(push(`${this.props.rootPath}/collections`));
         }
-    }
+    };
 
-    handleCollectionPageClick(uri) {
+    handleCollectionPageClick = uri => {
         if (uri === this.props.activePageURI) {
             return;
         }
@@ -368,9 +356,9 @@ export class CollectionDetailsController extends Component {
         this.props.dispatch(push(newURL));
 
         return newURL; //using 'return' so that we can test the correct new URL has been generated
-    }
+    };
 
-    async handleCollectionPageEditClick(page, state) {
+    handleCollectionPageEditClick = async (page, state) => {
         if (page.type === "dataset_details") {
             // This is a horrible hack to get the latest version url.
             // This could possibly be given to us from Zebedee.
@@ -425,7 +413,7 @@ export class CollectionDetailsController extends Component {
         window.location = newURL;
 
         return newURL; //return the URL so that we can test that the correct route is being used
-    }
+    };
 
     handleCollectionPageDeleteUndo(deleteTimer, uri, notificationID, pageType) {
         log.event("undo delete page from collection", log.data({ url: uri, type: pageType }));
@@ -456,7 +444,7 @@ export class CollectionDetailsController extends Component {
         return pageRoute; //using 'return' so that we can test the correct new URL has been generated
     }
 
-    handleCollectionPageDeleteClick(deletedPage, state) {
+    handleCollectionPageDeleteClick = (deletedPage, state) => {
         log.event(
             "deleting page from collection",
             log.data({
@@ -582,21 +570,21 @@ export class CollectionDetailsController extends Component {
         const notificationID = notifications.add(notification);
 
         return collectionURL; //using 'return' so that we can test the correct new URL has been generated
-    }
+    };
 
-    handleDrawerCloseClick() {
+    handleDrawerCloseClick = () => {
         this.setState({
             drawerIsAnimatable: true,
             drawerIsVisible: false
         });
         this.removeActiveCollectionGlobally();
-    }
+    };
 
-    handleRestoreDeletedContentClose() {
+    handleRestoreDeletedContentClose = () => {
         this.props.dispatch(push(url.resolve("../")));
-    }
+    };
 
-    handleRestoreMultiDeletedContentSuccess(updatedInProgressList) {
+    handleRestoreMultiDeletedContentSuccess = updatedInProgressList => {
         const mappedUpdatedInprogressList = updatedInProgressList.map(item => {
             return {
                 uri: item.uri,
@@ -612,9 +600,9 @@ export class CollectionDetailsController extends Component {
 
         this.props.dispatch(updatePagesInActiveCollection(updatedActiveCollection));
         this.handleRestoreDeletedContentClose();
-    }
+    };
 
-    handleRestoreSingleDeletedContentSuccess(restoredItem) {
+    handleRestoreSingleDeletedContentSuccess = restoredItem => {
         const addDeleteToInProgress = {
             uri: restoredItem.uri,
             title: restoredItem.title,
@@ -628,7 +616,7 @@ export class CollectionDetailsController extends Component {
 
         this.props.dispatch(updatePagesInActiveCollection(updatedActiveCollection));
         this.handleRestoreDeletedContentClose();
-    }
+    };
 
     renderLoadingCollectionDetails() {
         return (
