@@ -35,7 +35,7 @@ export class CreateDatasetTaxonomyController extends Component {
         };
     }
 
-    componentWillMount = () => {
+    UNSAFE_componentWillMount = () => {
         this.getTaxonomyNodes();
     };
 
