@@ -40,14 +40,9 @@ export class LoginController extends Component {
             requestPasswordChange: false,
             isSubmitting: false
         };
-
-        this.handleSubmit = this.handleSubmit.bind(this);
-        this.handleInputChange = this.handleInputChange.bind(this);
-        this.handlePasswordChangeCancel = this.handlePasswordChangeCancel.bind(this);
-        this.handlePasswordChangeSuccess = this.handlePasswordChangeSuccess.bind(this);
     }
 
-    componentWillMount() {
+    UNSAFE_componentWillMount() {
         if (this.props.isAuthenticated) {
             this.props.dispatch(push(`${this.props.rootPath}/collections`));
         }
@@ -177,7 +172,7 @@ export class LoginController extends Component {
             });
     }
 
-    handleSubmit(event) {
+    handleSubmit = event => {
         event.preventDefault();
 
         const credentials = {
@@ -188,9 +183,9 @@ export class LoginController extends Component {
         this.setState({ isSubmitting: true });
 
         this.requestLogin(credentials);
-    }
+    };
 
-    handleInputChange(event) {
+    handleInputChange = event => {
         const id = event.target.id;
         const value = event.target.value;
         const checked = event.target.checked;
@@ -235,24 +230,24 @@ export class LoginController extends Component {
                 notifications.add(notification);
             }
         }
-    }
+    };
 
-    handlePasswordChangeSuccess(newPassword) {
+    handlePasswordChangeSuccess = newPassword => {
         const credentials = {
             email: this.state.email.value,
             password: newPassword
         };
 
         this.requestLogin(credentials);
-    }
+    };
 
-    handlePasswordChangeCancel(event) {
+    handlePasswordChangeCancel = event => {
         event.preventDefault();
 
         this.setState({
             requestPasswordChange: false
         });
-    }
+    };
 
     render() {
         const inputs = [
