@@ -36,25 +36,22 @@ export default class Input extends Component {
             type: this.props.type,
             displayShowHide: (!this.props.disableShowPasswordText && this.props.type) === "password"
         };
-
-        this.showHide = this.showHide.bind(this);
-        this.moveCaretToEnd = this.moveCaretToEnd.bind(this);
     }
 
-    showHide(e) {
+    showHide = e => {
         e.preventDefault();
         e.stopPropagation();
         this.setState({
             type: this.state.type === "text" ? "password" : "text"
         });
-    }
+    };
 
-    moveCaretToEnd(event) {
+    moveCaretToEnd = event => {
         // Move caret to the end of the value on input focus
         const val = event.target.value;
         event.target.value = "";
         event.target.value = val;
-    }
+    };
 
     getFormClasses() {
         let formClasses = "form__input";
