@@ -1,4 +1,4 @@
-import React, { Component } from "react";
+import React from "react";
 import PropTypes from "prop-types";
 
 import NotificationItem from "./NotificationItem";
@@ -7,17 +7,15 @@ const propTypes = {
     notifications: PropTypes.arrayOf(PropTypes.object)
 };
 
-class Notifications extends Component {
-    render() {
-        return (
-            <ul className="notifications">
-                {this.props.notifications.map(notification => {
-                    return <NotificationItem {...notification} key={notification.id} />;
-                })}
-            </ul>
-        );
-    }
-}
+const Notifications = ({ notifications }) => {
+    return (
+        <ul className="notifications">
+            {notifications.map(notification => {
+                return <NotificationItem {...notification} key={notification.id} />;
+            })}
+        </ul>
+    );
+};
 
 Notifications.propTypes = propTypes;
 
